+++ gazebo/gui/model/ModelCreator.cc
@@ -85,6 +85,7 @@ ModelCreator::ModelCreator()
   this->jointMaker = new JointMaker();
 
   connect(g_editModelAct, SIGNAL(toggled(bool)), this, SLOT(OnEdit(bool)));
+
   this->inspectAct = new QAction(tr("Open Link Inspector"), this);
   connect(this->inspectAct, SIGNAL(triggered()), this,
       SLOT(OnOpenInspector()));
@@ -148,12 +149,20 @@ ModelCreator::ModelCreator()
        boost::bind(&ModelCreator::OnSetSelectedLink, this, _1, _2)));
 
   this->connections.push_back(
+     gui::model::Events::ConnectSetSelectedModelPlugin(
+       boost::bind(&ModelCreator::OnSetSelectedModelPlugin, this, _1, _2)));
+
+  this->connections.push_back(
       gui::Events::ConnectScaleEntity(
       boost::bind(&ModelCreator::OnEntityScaleChanged, this, _1, _2)));
 
   this->connections.push_back(
       gui::model::Events::ConnectShowLinkContextMenu(
       boost::bind(&ModelCreator::ShowContextMenu, this, _1)));
+
+  this->connections.push_back(
+      gui::model::Events::ConnectShowModelPluginContextMenu(
+      boost::bind(&ModelCreator::ShowModelPluginContextMenu, this, _1)));
 
   this->connections.push_back(
       gui::model::Events::ConnectRequestLinkRemoval(
@@ -1299,6 +1308,12 @@ void ModelCreator::RemoveLink(const std::string &_entity)
 }
 
 /////////////////////////////////////////////////
+void ModelCreator::OnRemoveModelPlugin(const QString &_name)
+{
+  this->RemoveModelPlugin(_name.toStdString());
+}
+
+/////////////////////////////////////////////////
 void ModelCreator::RemoveModelPlugin(const std::string &_name)
 {
   boost::recursive_mutex::scoped_lock lock(*this->updateMutex);
@@ -1335,6 +1350,13 @@ bool ModelCreator::OnKeyPress(const common::KeyEvent &_event)
         this->OnDelete(linkVis->GetName());
       }
       this->DeselectAll();
+    }
+    else if (!this->selectedModelPlugins.empty())
+    {
+      for (auto plugin : this->selectedModelPlugins)
+      {
+        this->RemoveModelPlugin(plugin);
+      }
     }
   }
   else if (_event.control)
@@ -1515,6 +1537,45 @@ void ModelCreator::ShowContextMenu(const std::string &_link)
   }
   QAction *deleteAct = new QAction(tr("Delete"), this);
   connect(deleteAct, SIGNAL(triggered()), this, SLOT(OnDelete()));
+  menu.addAction(deleteAct);
+
+  menu.exec(QCursor::pos());
+}
+
+/////////////////////////////////////////////////
+void ModelCreator::ShowModelPluginContextMenu(const std::string &_name)
+{
+  auto it = this->allModelPlugins.find(_name);
+  if (it == this->allModelPlugins.end())
+    return;
+
+  // Open inspector
+  QAction *inspectorAct = new QAction(tr("Open Model Plugin Inspector"), this);
+
+  // Map signals to pass argument
+  QSignalMapper *inspectorMapper = new QSignalMapper(this);
+
+  connect(inspectorAct, SIGNAL(triggered()), inspectorMapper, SLOT(map()));
+  inspectorMapper->setMapping(inspectorAct, QString::fromStdString(_name));
+
+  connect(inspectorMapper, SIGNAL(mapped(QString)), this,
+      SLOT(OnOpenModelPluginInspector(QString)));
+
+  // Delete
+  QAction *deleteAct = new QAction(tr("Delete"), this);
+
+  // Map signals to pass argument
+  QSignalMapper *deleteMapper = new QSignalMapper(this);
+
+  connect(deleteAct, SIGNAL(triggered()), deleteMapper, SLOT(map()));
+  deleteMapper->setMapping(deleteAct, QString::fromStdString(_name));
+
+  connect(deleteMapper, SIGNAL(mapped(QString)), this,
+      SLOT(OnRemoveModelPlugin(QString)));
+
+  // Menu
+  QMenu menu;
+  menu.addAction(inspectorAct);
   menu.addAction(deleteAct);
 
   menu.exec(QCursor::pos());
@@ -1907,6 +1968,25 @@ void ModelCreator::OnSetSelectedLink(const std::string &_name,
   this->SetSelected(_name, _selected);
 }
 
+/////////////////////////////////////////////////
+void ModelCreator::OnSetSelectedModelPlugin(const std::string &_name,
+    const bool _selected)
+{
+  auto plugin = this->allModelPlugins.find(_name);
+  if (plugin == this->allModelPlugins.end())
+    return;
+
+  auto it = std::find(this->selectedModelPlugins.begin(),
+      this->selectedModelPlugins.end(), _name);
+  if (_selected && it == this->selectedModelPlugins.end())
+  {
+    this->selectedModelPlugins.push_back(_name);
+  }
+  else if (!_selected && it != this->selectedModelPlugins.end())
+  {
+    this->selectedModelPlugins.erase(it);
+  }
+}
 
 /////////////////////////////////////////////////
 void ModelCreator::ModelChanged()
@@ -2022,6 +2102,12 @@ void ModelCreator::AddModelPlugin(const sdf::ElementPtr _pluginElem)
     // Notify addition
     gui::model::Events::modelPluginInserted(name);
   }
+}
+
+/////////////////////////////////////////////////
+void ModelCreator::OnOpenModelPluginInspector(const QString &_name)
+{
+  this->OpenModelPluginInspector(_name.toStdString());
 }
 
 /////////////////////////////////////////////////